Add templated render method to callback events
The major use case (for me) for this app is to convert the JSON returned
from Trello into a message and then send it somewhere else. The new
render method uses the django templating to render the JSON payload as
the template context.

The templates are in /templates/trello_webhooks/<event_type>.html, and
can be overridden by other apps, so that users have full control over
the output. The default templates are very functional, so should be
treated as samples only.