Display image attachments in HC as inline images
This implements the above feature as described in
[Issue #2](https://github.com/yunojuno/django-trello-webhooks/issues/2).

Type support
------------
I couldn’t find a comprehensive list of what image types are supported
for inline display in HC, so I added support for standard image content-types (`image/...`).
Gifv is supported in HC but is not really an image.

Type detection
--------------
Attachments in the Trello API appear to have a `mimeType` field.
However in my tests it failed to work reliably (was usually null
for simple jpegs) so I didn’t use it.

There is no 100% valid solution for detecting the real type of the file
without actually inspecting the contents; the extension and mime type
can be faked easily. For the purpose of this feature it’s an overkill
to actually download and open a file, so it will just assume the mime type
from the extension. Failing that, it will use whatever is reported
by a head request on its hosting server.