Add get_attachment_type method to CallbackEvent.
Added get_attachment_type method to Callbackevent. We are using requests
to obtain the initial chunk of the file (preventing downloading the
entire attachment). New dependencies were added - requests and
python-magic which helps to get the mime type. My assumption is that the
attachment's mime type signature is within the first 1024 bytes fo the
file (seems to be the case for almost all files but .iso).