Add test for get_attachment_type.
Added tests for get_attachment_type method in test_models.py.
Had to mock out response object and file object to avoid relying on the
internet connection. This obviously means that the attachment types are
fed manually through use of mocks and we only test the logic of the
get_attachment_type method not the actual working of the
get / from_buffer.